@@ -1,128 +1,68 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
 	<artifactId>biosdk-services</artifactId>
 	<groupId>io.mosip.biosdk</groupId>
 	<version>1.2.1-SNAPSHOT</version>
 	<name>biosdk-services</name>
-
-	<licenses>
-		<license>
-			<name>MPL 2.0</name>
-			<url>https://www.mozilla.org/en-US/MPL/2.0/</url>
-		</license>
-	</licenses>
-
-	<scm>
-		<connection>scm:git:git://github.com/mosip/biosdk-services.git</connection>
-		<developerConnection>scm:git:ssh://github.com:mosip/biosdk-services.git</developerConnection>
-		<url>https://github.com/mosip/biosdk-services</url>
-		<tag>HEAD</tag>
-	</scm>
-
-	<developers>
-		<developer>
-			<name>Mosip</name>
-			<email>mosip.emailnotifier@gmail.com</email>
-			<organization>io.mosip</organization>
-			<organizationUrl>https://mosip.io</organizationUrl>
-		</developer>
-	</developers>
-	<repositories>
-        	<repository>
-            		<id>central</id>
-            		<name>MavenCentral</name>
-            		<layout>default</layout>
-            		<url>https://repo1.maven.org/maven2</url>
-            		<snapshots>
-            			<enabled>false</enabled>
-            		</snapshots>
-        	</repository>
-        	<repository>
-            		<id>ossrh</id>
-            		<name>MavenSnapshot</name>
-            		<url>https://oss.sonatype.org/content/repositories/snapshots</url>
-            		<layout>default</layout>
-            		<snapshots>
-            		<enabled>true</enabled>
-            		</snapshots>
-        	</repository>
-    	</repositories>
 	<description>Sample implementation of biometrics SDK services</description>
 	<url>https://github.com/mosip/biosdk-services</url>
 
 	<properties>
-		<java.version>11</java.version>
-		<!-- maven -->
-		<maven.compiler.source>11</maven.compiler.source>
-		<maven.compiler.target>11</maven.compiler.target>
-		<!-- Core -->
-		<spring.boot.version>2.0.2.RELEASE</spring.boot.version>
-		<spring-cloud-config.version>2.0.2.RELEASE</spring-cloud-config.version>
-		<maven.surefire.plugin.version>2.22.0</maven.surefire.plugin.version>
-		<maven.jacoco.plugin.version>0.8.5</maven.jacoco.plugin.version>
-		<maven.sonar.plugin.version>3.7.0.1746</maven.sonar.plugin.version>
+		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+		<java.version>21</java.version>
 
-		<!-- Lombok -->
-		<lombok.version>1.18.8</lombok.version>
+		<!-- maven -->
+		<maven.compiler.source>${java.version}</maven.compiler.source>
+		<maven.compiler.target>${java.version}</maven.compiler.target>
+		<maven.compiler.version>3.11.0</maven.compiler.version>
+		<maven.jar.plugin.version>3.0.2</maven.jar.plugin.version>
+		<maven.war.plugin.version>3.1.0</maven.war.plugin.version>
+		<maven.javadoc.plugin.version>3.2.0</maven.javadoc.plugin.version>
+		<maven.shade.plugin.version>2.3</maven.shade.plugin.version>
+		<jacoco.maven.plugin.version>0.8.11</jacoco.maven.plugin.version>
+		<maven.deploy.plugin.version>3.1.1</maven.deploy.plugin.version>
+		<maven.source.plugin.version>3.3.1</maven.source.plugin.version>
+		<maven.gpg.plugin.version>3.2.3</maven.gpg.plugin.version>
+		<maven.jar.plugin.version>3.0.2</maven.jar.plugin.version>
+		<maven.war.plugin.version>3.1.0</maven.war.plugin.version>
+		<maven.assembly.plugin.version>3.3.0</maven.assembly.plugin.version>
+		<maven.surefire.plugin.version>3.1.2</maven.surefire.plugin.version>
+		<spring.boot.maven.plugin.version>3.2.5</spring.boot.maven.plugin.version>
+		<nexus.staging.maven.plugin.version>1.6.7</nexus.staging.maven.plugin.version>
+		<maven.sonar.plugin.version>3.7.0.1746</maven.sonar.plugin.version>
+		<!-- git -->
+		<git.commit.id.plugin.version>3.0.1</git.commit.id.plugin.version>
 
 		<!-- Swagger -->
-		<swagger.version>2.9.2</swagger.version>
-
+		<springdoc.openapi.starter.webmvc.ui.version>2.5.0</springdoc.openapi.starter.webmvc.ui.version>
 
 		<!-- Mosip kernel -->
+		<kernel.bom.version>1.2.1-SNAPSHOT</kernel.bom.version>
+		<kernel.core.version>1.2.1-SNAPSHOT</kernel.core.version>
 		<kernel.biometrics.api.version>1.2.1-SNAPSHOT</kernel.biometrics.api.version>
 		<kernel.logger.logback.version>1.2.1-SNAPSHOT</kernel.logger.logback.version>
-		<kernel.core.version>1.2.1-SNAPSHOT</kernel.core.version>
-		<!-- Spring -->
-		<spring.boot.version>2.0.2.RELEASE</spring.boot.version>
 	</properties>
 
 	<dependencyManagement>
 		<dependencies>
 			<dependency>
-				<groupId>org.springframework.boot</groupId>
-				<artifactId>spring-boot-dependencies</artifactId>
-				<version>${spring.boot.version}</version>
+				<groupId>io.mosip.kernel</groupId>
+				<artifactId>kernel-bom</artifactId>
+				<version>${kernel.bom.version}</version>
 				<type>pom</type>
 				<scope>import</scope>
-				<exclusions>
-					<exclusion>
-						<groupId>org.springframework.boot</groupId>
-						<artifactId>spring-boot-starter-json</artifactId>
-					</exclusion>
-				</exclusions>
-			</dependency>
-			<dependency>
-				<groupId>com.fasterxml.jackson.core</groupId>
-				<artifactId>jackson-databind</artifactId>
-				<version>2.11.0</version>
-			</dependency>
-			<dependency>
-				<groupId>com.fasterxml.jackson.core</groupId>
-				<artifactId>jackson-core</artifactId>
-				<version>2.11.0</version>
 			</dependency>
 		</dependencies>
 	</dependencyManagement>
 
-
 	<dependencies>
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-starter-config</artifactId>
-			<version>${spring-cloud-config.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-devtools</artifactId>
-			<scope>runtime</scope>
-		</dependency>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-web</artifactId>
-			<version>${spring.boot.version}</version>
 			<exclusions>
 				<exclusion>
 					<groupId>org.springframework.boot</groupId>
