@@ -1,20 +1,33 @@
 package io.mosip.biosdk.services.dto;
 
-import lombok.Getter;
+import lombok.AllArgsConstructor;
+import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 import lombok.ToString;
 
-@Getter
-@Setter
+/**
+ * Data Transfer Object (DTO) for encapsulating error details in the MOSIP
+ * Biometric SDK service. This class holds an error code and an error message.
+ * <p>
+ * The {@code ErrorDto} class uses Lombok annotations for boilerplate code
+ * reduction.
+ * </p>
+ *
+ *
+ * @since 1.0.0
+ */
+@Data
 @NoArgsConstructor
+@AllArgsConstructor
 @ToString
 public class ErrorDto {
-    private String code;
-    private String message;
+	/**
+	 * The error code.
+	 */
+	private String code;
 
-    public ErrorDto(String code, String message){
-        this.code = code;
-        this.message = message;
-    }
+	/**
+	 * The error message.
+	 */
+	private String message;
 }
