@@ -10,11 +10,10 @@ import java.util.List;
 import java.util.function.BiConsumer;
 
 import org.apache.commons.codec.digest.DigestUtils;
-import org.json.simple.parser.ParseException;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
 
 import io.mosip.biosdk.services.dto.RequestDto;
 import io.mosip.biosdk.services.impl.spec_1_0.dto.request.CheckQualityRequestDto;
@@ -29,38 +28,103 @@ import io.mosip.kernel.biometrics.entities.BIRInfo;
 import io.mosip.kernel.biometrics.entities.BiometricRecord;
 import io.mosip.kernel.core.util.DateUtils;
 
+/**
+ * Utility class providing methods to convert various DTOs (Data Transfer
+ * Objects) to their JSON string representations.
+ * <p>
+ * This class uses Gson for JSON serialization and provides methods to convert
+ * different types of DTOs including {@link ExtractTemplateRequestDto},
+ * {@link MatchRequestDto}, {@link InitRequestDto},
+ * {@link CheckQualityRequestDto}, {@link SegmentRequestDto},
+ * {@link ConvertFormatRequestDto}, {@link BDBInfo}, {@link BIRInfo}, and
+ * {@link BiometricRecord}.
+ * </p>
+ * <p>
+ * The conversion methods handle null values gracefully and use helper methods
+ * to format dates and perform hashing operations on byte arrays.
+ * </p>
+ * <p>
+ * This class is designed to assist in logging, debugging, or transmitting data
+ * in a structured JSON format compatible with the MOSIP Bio SDK services.
+ * </p>
+ * 
+ * @since 1.0.0
+ */
 @Component
 public class Utils {
-    @Autowired
-    private Gson gson;
+	private Gson gson;
 
-    private String utcDateTimePattern = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'";
+	private String utcDateTimePattern = "yyyy-MM-dd'T'HH:mm:ss.SSS'Z'";
+	private static final String FLAGS = ", \"flags\":";
+	private static final String SAMPLE = ", \"sample\":";
 
-    public String getCurrentResponseTime() {
-        return DateUtils.formatDate(new Date(System.currentTimeMillis()), utcDateTimePattern);
-    }
+	/**
+	 * Constructs a new Utils instance initializing Gson for JSON serialization.
+	 */
+	public Utils() {
+		gson = new GsonBuilder().serializeNulls().create();
+	}
 
-    public RequestDto getRequestInfo(String request) throws ParseException {
-        return gson.fromJson(request, RequestDto.class);
-    }
+	/**
+	 * Retrieves the current timestamp formatted as per UTC date-time pattern.
+	 *
+	 * @return formatted current timestamp.
+	 */
+	public String getCurrentResponseTime() {
+		return DateUtils.formatDate(new Date(System.currentTimeMillis()), utcDateTimePattern);
+	}
 
-	public static String base64Decode(String data){
-        return new String(Base64.getDecoder().decode(data), StandardCharsets.UTF_8);
-    }
-	
+	/**
+	 * Parses the provided JSON string into a RequestDto object.
+	 *
+	 * @param request JSON string representation of the request.
+	 * @return RequestDto object parsed from the JSON.
+	 */
+	public RequestDto getRequestInfo(String request) {
+		return gson.fromJson(request, RequestDto.class);
+	}
+
+	/**
+	 * Decodes a Base64 encoded string into its original UTF-8 string
+	 * representation.
+	 *
+	 * @param data Base64 encoded string to decode.
+	 * @return decoded UTF-8 string.
+	 */
+	public static String base64Decode(String data) {
+		return new String(Base64.getDecoder().decode(data), StandardCharsets.UTF_8);
+	}
+
+	/**
+	 * Converts a BiometricRecord object to its string representation. If the
+	 * BiometricRecord object is null, returns "null". Otherwise, constructs a
+	 * JSON-like string representation including key attributes such as biometric
+	 * information, version, and segments.
+	 *
+	 * @param biometricRecord The BiometricRecord object to convert to string.
+	 * @return String representation of the BiometricRecord object.
+	 */
 	public String toString(BiometricRecord biometricRecord) {
-		if(biometricRecord == null) {
+		if (biometricRecord == null) {
 			return "null";
 		}
-		
+
 		StringBuilder stringBuilder = new StringBuilder();
 		appendString(biometricRecord, stringBuilder);
 		return stringBuilder.toString();
 	}
 
-    private void appendString(BiometricRecord biometricRecord, StringBuilder stringBuilder) {
-    	if(biometricRecord == null) {
-    		stringBuilder.append("null");
+	/**
+	 * Helper method to append the string representation of a BiometricRecord object
+	 * to StringBuilder. Constructs a JSON-like format including biometric
+	 * information, version, and segments.
+	 *
+	 * @param biometricRecord The BiometricRecord object to convert to string.
+	 * @param stringBuilder   StringBuilder to append the string representation.
+	 */
+	private void appendString(BiometricRecord biometricRecord, StringBuilder stringBuilder) {
+		if (biometricRecord == null) {
+			stringBuilder.append("null");
 		} else {
 			stringBuilder.append("{");
 			stringBuilder.append(" \"_modelClass\": \"BiometricRecord\"");
@@ -72,34 +136,61 @@ public class Utils {
 			stringBuilder.append(stringOf(biometricRecord.getVersion()));
 			stringBuilder.append(", \"segments\":");
 			List<BIR> segments = biometricRecord.getSegments();
-			if(segments == null) {
-	    		stringBuilder.append("null");
+			if (segments == null) {
+				stringBuilder.append("null");
 			} else {
 				appendString(segments.stream().iterator(), stringBuilder, this::appendString);
 			}
 			stringBuilder.append(" }");
 		}
 	}
-    
-    private String stringOf(Object obj) {
-    	return obj == null ? "null" : gson.toJson(obj);
-    }
 
-	private <T> void appendString(Iterator<T> iterator, StringBuilder stringBuilder, BiConsumer<T, StringBuilder> appendBiConsumer) {
+	/**
+	 * Converts the provided object to its JSON representation using Gson. If the
+	 * object is {@code null}, returns the string {@code "null"}.
+	 *
+	 * @param obj The object to convert to JSON.
+	 * @return JSON representation of the object or {@code "null"} if the object is
+	 *         null.
+	 */
+	private String stringOf(Object obj) {
+		return obj == null ? "null" : gson.toJson(obj);
+	}
+
+	/**
+	 * Appends elements from the provided iterator to the StringBuilder using a
+	 * custom BiConsumer for element-specific appending logic.
+	 *
+	 * @param iterator         The iterator containing elements to append.
+	 * @param stringBuilder    The StringBuilder to append elements to.
+	 * @param appendBiConsumer The BiConsumer defining how to append each element to
+	 *                         the StringBuilder. Typically encapsulates specific
+	 *                         logic for each element type.
+	 */
+	private <T> void appendString(Iterator<T> iterator, StringBuilder stringBuilder,
+			BiConsumer<T, StringBuilder> appendBiConsumer) {
 		stringBuilder.append("[ ");
 		while (iterator.hasNext()) {
 			T element = iterator.next();
 			appendBiConsumer.accept(element, stringBuilder);
-			if(iterator.hasNext()) {
+			if (iterator.hasNext()) {
 				stringBuilder.append(", ");
 			}
 		}
 		stringBuilder.append(" ]");
 	}
 
+	/**
+	 * Appends a Biometric Information Record (BIR) object to the StringBuilder in
+	 * JSON-like format, including its various attributes and hashes. If the
+	 * provided BIR object is null, appends the string "null".
+	 *
+	 * @param bir           The Biometric Information Record (BIR) object to append.
+	 * @param stringBuilder The StringBuilder to which the BIR object is appended.
+	 */
 	private void appendString(BIR bir, StringBuilder stringBuilder) {
-		if(bir == null) {
-    		stringBuilder.append("null");
+		if (bir == null) {
+			stringBuilder.append("null");
 		} else {
 			stringBuilder.append("{");
 			stringBuilder.append(" \"_modelClass\": \"BIR\"");
@@ -123,39 +214,61 @@ public class Utils {
 		}
 	}
 
+	/**
+	 * Computes the SHA-256 hash of the provided byte array and returns it as a
+	 * formatted string. If the byte array is null, returns the string "null".
+	 *
+	 * @param byteArray The byte array to compute the SHA-256 hash from.
+	 * @return A formatted string representing the SHA-256 hash of the byte array.
+	 */
 	private static String getHashOfBytes(byte[] byteArray) {
-		return byteArray == null ? "null" : "\""+ DigestUtils.sha256Hex(byteArray) + "\"";
+		return byteArray == null ? "null" : "\"" + DigestUtils.sha256Hex(byteArray) + "\"";
 	}
-	
+
+	/**
+	 * Returns a JSON string representation of the ExtractTemplateRequestDto object.
+	 * If the provided object is null, returns the string "null".
+	 *
+	 * @param extractTemplateRequestDto The ExtractTemplateRequestDto object to
+	 *                                  convert to JSON string.
+	 * @return JSON string representation of the ExtractTemplateRequestDto object.
+	 */
 	public String toString(ExtractTemplateRequestDto extractTemplateRequestDto) {
-		if(extractTemplateRequestDto == null) {
+		if (extractTemplateRequestDto == null) {
 			return "null";
 		}
 		StringBuilder stringBuilder = new StringBuilder();
 		stringBuilder.append("{");
 		stringBuilder.append(" \"_modelClass\": \"ExtractTemplateRequestDto\"");
-		stringBuilder.append(", \"flags\":");
+		stringBuilder.append(FLAGS);
 		stringBuilder.append(stringOf(extractTemplateRequestDto.getFlags()));
 		stringBuilder.append(", \"modalitiesToExtract\": ");
 		stringBuilder.append(stringOf(extractTemplateRequestDto.getModalitiesToExtract()));
-		stringBuilder.append(", \"sample\": ");
+		stringBuilder.append(SAMPLE);
 		appendString(extractTemplateRequestDto.getSample(), stringBuilder);
 		stringBuilder.append(" }");
 		return stringBuilder.toString();
 	}
 
+	/**
+	 * Returns a JSON string representation of the MatchRequestDto object. If the
+	 * provided object is null, returns the string "null".
+	 *
+	 * @param matchRequestDto The MatchRequestDto object to convert to JSON string.
+	 * @return JSON string representation of the MatchRequestDto object.
+	 */
 	public String toString(MatchRequestDto matchRequestDto) {
-		if(matchRequestDto == null) {
+		if (matchRequestDto == null) {
 			return "null";
 		}
 		StringBuilder stringBuilder = new StringBuilder();
 		stringBuilder.append("{");
 		stringBuilder.append(" \"_modelClass\": \"MatchRequestDto\"");
-		stringBuilder.append(", \"flags\":");
+		stringBuilder.append(", " + FLAGS + ":");
 		stringBuilder.append(stringOf(matchRequestDto.getFlags()));
 		stringBuilder.append(", \"modalitiesToMatch\": ");
 		stringBuilder.append(stringOf(matchRequestDto.getModalitiesToMatch()));
-		stringBuilder.append(", \"sample\": ");
+		stringBuilder.append(SAMPLE);
 		appendString(matchRequestDto.getSample(), stringBuilder);
 		stringBuilder.append(", \"gallery\": ");
 		appendString(Arrays.stream(matchRequestDto.getGallery()).iterator(), stringBuilder, this::appendString);
@@ -163,8 +276,15 @@ public class Utils {
 		return stringBuilder.toString();
 	}
 
+	/**
+	 * Returns a JSON string representation of the InitRequestDto object. If the
+	 * provided object is null, returns the string "null".
+	 *
+	 * @param initRequestDto The InitRequestDto object to convert to JSON string.
+	 * @return JSON string representation of the InitRequestDto object.
+	 */
 	public String toString(InitRequestDto initRequestDto) {
-		if(initRequestDto == null) {
+		if (initRequestDto == null) {
 			return "null";
 		}
 		StringBuilder stringBuilder = new StringBuilder();
@@ -176,42 +296,66 @@ public class Utils {
 		return stringBuilder.toString();
 	}
 
+	/**
+	 * Returns a JSON string representation of the CheckQualityRequestDto object. If
+	 * the provided object is null, returns the string "null".
+	 *
+	 * @param checkQualityRequestDto The CheckQualityRequestDto object to convert to
+	 *                               JSON string.
+	 * @return JSON string representation of the CheckQualityRequestDto object.
+	 */
 	public String toString(CheckQualityRequestDto checkQualityRequestDto) {
-		if(checkQualityRequestDto == null) {
+		if (checkQualityRequestDto == null) {
 			return "null";
 		}
 		StringBuilder stringBuilder = new StringBuilder();
 		stringBuilder.append("{");
 		stringBuilder.append(" \"_modelClass\": \"CheckQualityRequestDto\"");
-		stringBuilder.append(", \"flags\":");
+		stringBuilder.append(FLAGS);
 		stringBuilder.append(stringOf(checkQualityRequestDto.getFlags()));
 		stringBuilder.append(", \"modalitiesToCheck\": ");
 		stringBuilder.append(stringOf(checkQualityRequestDto.getModalitiesToCheck()));
-		stringBuilder.append(", \"sample\": ");
+		stringBuilder.append(SAMPLE);
 		appendString(checkQualityRequestDto.getSample(), stringBuilder);
 		stringBuilder.append(" }");
 		return stringBuilder.toString();
 	}
 
+	/**
+	 * Returns a JSON string representation of the SegmentRequestDto object. If the
+	 * provided object is null, returns the string "null".
+	 *
+	 * @param segmentRequestDto The SegmentRequestDto object to convert to JSON
+	 *                          string.
+	 * @return JSON string representation of the SegmentRequestDto object.
+	 */
 	public String toString(SegmentRequestDto segmentRequestDto) {
-		if(segmentRequestDto == null) {
+		if (segmentRequestDto == null) {
 			return "null";
 		}
 		StringBuilder stringBuilder = new StringBuilder();
 		stringBuilder.append("{");
 		stringBuilder.append(" \"_modelClass\": \"SegmentRequestDto\"");
-		stringBuilder.append(", \"flags\":");
+		stringBuilder.append(FLAGS);
 		stringBuilder.append(stringOf(segmentRequestDto.getFlags()));
 		stringBuilder.append(", \"modalitiesToSegment\": ");
 		stringBuilder.append(stringOf(segmentRequestDto.getModalitiesToSegment()));
-		stringBuilder.append(", \"sample\": ");
+		stringBuilder.append(SAMPLE);
 		appendString(segmentRequestDto.getSample(), stringBuilder);
 		stringBuilder.append(" }");
 		return stringBuilder.toString();
 	}
 
+	/**
+	 * Returns a JSON string representation of the ConvertFormatRequestDto object.
+	 * If the provided object is null, returns the string "null".
+	 *
+	 * @param convertFormatRequestDto The ConvertFormatRequestDto object to convert
+	 *                                to JSON string.
+	 * @return JSON string representation of the ConvertFormatRequestDto object.
+	 */
 	public String toString(ConvertFormatRequestDto convertFormatRequestDto) {
-		if(convertFormatRequestDto == null) {
+		if (convertFormatRequestDto == null) {
 			return "null";
 		}
 		StringBuilder stringBuilder = new StringBuilder();
@@ -223,7 +367,7 @@ public class Utils {
 		stringBuilder.append(stringOf(convertFormatRequestDto.getTargetFormat()));
 		stringBuilder.append(", \"modalitiesToConvert\": ");
 		stringBuilder.append(stringOf(convertFormatRequestDto.getModalitiesToConvert()));
-		stringBuilder.append(", \"sample\": ");
+		stringBuilder.append(SAMPLE);
 		appendString(convertFormatRequestDto.getSample(), stringBuilder);
 		stringBuilder.append(", \"sourceParams\":");
 		stringBuilder.append(stringOf(convertFormatRequestDto.getSourceParams()));
@@ -233,9 +377,16 @@ public class Utils {
 		stringBuilder.append(" }");
 		return stringBuilder.toString();
 	}
-	
+
+	/**
+	 * Returns a JSON string representation of the BDBInfo object. If the provided
+	 * object is null, returns the string "null".
+	 *
+	 * @param bdbInfo The BDBInfo object to convert to JSON string.
+	 * @return JSON string representation of the BDBInfo object.
+	 */
 	public String toString(BDBInfo bdbInfo) {
-		if(bdbInfo == null) {
+		if (bdbInfo == null) {
 			return "null";
 		}
 		StringBuilder stringBuilder = new StringBuilder();
@@ -278,9 +429,16 @@ public class Utils {
 		stringBuilder.append(" }");
 		return stringBuilder.toString();
 	}
-	
+
+	/**
+	 * Returns a JSON string representation of the BIRInfo object. If the provided
+	 * object is null, returns the string "null".
+	 *
+	 * @param birInfo The BIRInfo object to convert to JSON string.
+	 * @return JSON string representation of the BIRInfo object.
+	 */
 	public String toString(BIRInfo birInfo) {
-		if(birInfo == null) {
+		if (birInfo == null) {
 			return "null";
 		}
 		StringBuilder stringBuilder = new StringBuilder();
@@ -304,10 +462,24 @@ public class Utils {
 		return stringBuilder.toString();
 	}
 
+	/**
+	 * Converts a LocalDateTime object to its ISO-8601 string representation. If the
+	 * provided LocalDateTime object is null, returns the string "null".
+	 *
+	 * @param localDateTime The LocalDateTime object to convert.
+	 * @return ISO-8601 string representation of the LocalDateTime object.
+	 */
 	private String dateAsString(LocalDateTime localDateTime) {
 		return localDateTime == null ? "null" : DateUtils.formatToISOString(localDateTime);
 	}
 
+	/**
+	 * Converts a Boolean object to its string representation. If the provided
+	 * Boolean object is null, returns the string "null".
+	 *
+	 * @param bool The Boolean object to convert.
+	 * @return String representation of the Boolean object.
+	 */
 	private static String booleanAsString(Boolean bool) {
 		return bool == null ? "null" : Boolean.toString(bool);
 	}
